code/requirements.txt
matplotlib==3.5.1
numpy==1.21.5
opencv_python==4.5.4.60
pandas==1.2.1
Pillow==9.1.0
scikit_learn==1.0.2
seaborn==0.11.2
torch==1.11.0+cu113
torchvision==0.12.0+cu113
